refactor/AB#74466_ABC-remove-duplicated-js-and-css-code-from-the-frontend-repo-code refactor: start refactoring graphql repeated queries, mutations and subscriptions across the safe library refactor: added shared config for pipeline folder spec and stories files feat: added repeated code report for libs folder(would be deleted after all work is done)
