+++ libs/safe/src/lib/components/grid-layout/add-layout-modal/add-layout-modal.component.ts
@@ -1,14 +1,12 @@
 import { Component, Inject, OnInit, ViewChild } from '@angular/core';
 import { SafeGridLayoutService } from '../../../services/grid-layout/grid-layout.service';
 import { Form } from '../../../models/form.model';
-import { Resource } from '../../../models/resource.model';
-import { Apollo, QueryRef } from 'apollo-angular';
 import {
-  GetResourceLayoutsResponse,
-  GET_RESOURCE_LAYOUTS,
-  GetFormLayoutsResponse,
-  GET_FORM_LAYOUTS,
-} from './graphql/queries';
+  Resource,
+  ResourceQueryResponse,
+} from '../../../models/resource.model';
+import { Apollo, QueryRef } from 'apollo-angular';
+import { GetFormLayoutsResponse, GET_FORM_LAYOUTS } from './graphql/queries';
 import { UntypedFormControl } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
@@ -22,6 +20,7 @@ import {
 import { ButtonModule } from '@oort-front/ui';
 import { takeUntil } from 'rxjs';
 import { SafeUnsubscribeComponent } from '../../utils/unsubscribe/unsubscribe.component';
+import { GET_RESOURCE_LAYOUTS_FULL } from '../../../graphql/queries/resource-queries';
 
 /**
  * Data needed for the dialog, should contain a layouts array, a form and a resource
@@ -60,7 +59,7 @@ export class AddLayoutModalComponent
   public hasLayouts = false;
   public nextStep = false;
   public queryRef!:
-    | QueryRef<GetResourceLayoutsResponse>
+    | QueryRef<ResourceQueryResponse>
     | QueryRef<GetFormLayoutsResponse>
     | null;
   public selectedLayoutControl = new UntypedFormControl('');
@@ -93,8 +92,8 @@ export class AddLayoutModalComponent
 
   ngOnInit() {
     if (this.resource)
-      this.queryRef = this.apollo.watchQuery<GetResourceLayoutsResponse>({
-        query: GET_RESOURCE_LAYOUTS,
+      this.queryRef = this.apollo.watchQuery<ResourceQueryResponse>({
+        query: GET_RESOURCE_LAYOUTS_FULL,
         variables: {
           resource: this.resource?.id,
         },
