+++ libs/safe/src/lib/components/ui/aggregation-builder/pipeline/add-field-stage/add-field-stage.component.spec.ts
@@ -1,31 +1,13 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { UntypedFormArray } from '@angular/forms';
-import {
-  TranslateModule,
-  TranslateService,
-  TranslateFakeLoader,
-  TranslateLoader,
-} from '@ngx-translate/core';
 import { SafeAddFieldStageComponent } from './add-field-stage.component';
+import { setupSpecConfig } from '../shared-config/spec/config';
 
 describe('SafeAddFieldStageComponent', () => {
   let component: SafeAddFieldStageComponent;
   let fixture: ComponentFixture<SafeAddFieldStageComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      providers: [TranslateService],
-      declarations: [SafeAddFieldStageComponent],
-      imports: [
-        TranslateModule.forRoot({
-          loader: {
-            provide: TranslateLoader,
-            useClass: TranslateFakeLoader,
-          },
-        }),
-      ],
-    }).compileComponents();
-  });
+  beforeEach(async () => setupSpecConfig(SafeAddFieldStageComponent));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(SafeAddFieldStageComponent);
