+++ libs/safe/src/lib/components/ui/aggregation-builder/pipeline/add-field-stage/add-field-stage.stories.ts
@@ -1,65 +1,17 @@
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { Meta, moduleMetadata, StoryFn } from '@storybook/angular';
+import { Meta, StoryFn } from '@storybook/angular';
 import { SafeAddFieldStageComponent } from './add-field-stage.component';
-import { SafePipelineModule } from '../pipeline.module';
-import { StorybookTranslateModule } from '../../../../storybook-translate/storybook-translate-module';
-import { UntypedFormBuilder, ReactiveFormsModule } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 import { Accumulators } from '../expressions/operators';
+import {
+  DEFAULT_FIELDS,
+  SHARED_PIPELINE_STORIES_CONFIG,
+} from '../shared-config/stories/config';
 
 export default {
   component: SafeAddFieldStageComponent,
-  decorators: [
-    moduleMetadata({
-      imports: [
-        SafePipelineModule,
-        BrowserAnimationsModule,
-        StorybookTranslateModule,
-        ReactiveFormsModule,
-      ],
-      providers: [
-        {
-          provide: 'environment',
-          useValue: {},
-        },
-      ],
-    }),
-  ],
+  ...SHARED_PIPELINE_STORIES_CONFIG,
   title: 'UI/Aggregation builder/Stages/Add fields',
-  args: {
-    currentForms: [],
-    filteredForms: [],
-  },
 } as Meta;
-
-/** Defining the default fields that are available in the storybook */
-const DEFAULT_FIELDS = [
-  {
-    name: 'date',
-    args: [],
-    type: {
-      name: 'Date',
-      kind: 'SCALAR',
-      ofType: null,
-    },
-  },
-  {
-    name: 'description',
-    args: [],
-    type: {
-      name: 'String',
-      kind: 'SCALAR',
-      ofType: null,
-    },
-  },
-  {
-    name: 'status',
-    type: {
-      name: 'String',
-      kind: 'SCALAR',
-      ofType: null,
-    },
-  },
-];
 
 /** Creating a new instance of the FormBuilder class. */
 const fb = new UntypedFormBuilder();
