+++ libs/safe/src/lib/components/widgets/editor/editor.component.ts
@@ -6,7 +6,6 @@ import {
   GET_LAYOUT,
   GET_RESOURCE_METADATA,
   GetLayoutQueryResponse,
-  GetResourceMetadataQueryResponse,
 } from '../summary-card/graphql/queries';
 import { clone, get } from 'lodash';
 import { QueryBuilderService } from '../../../services/query-builder/query-builder.service';
@@ -14,6 +13,7 @@ import { DataTemplateService } from '../../../services/data-template/data-templa
 import { Dialog } from '@angular/cdk/dialog';
 import { SnackbarService } from '@oort-front/ui';
 import { TranslateService } from '@ngx-translate/core';
+import { ResourceQueryResponse } from '../../../models/resource.model';
 
 /**
  * Text widget component using KendoUI
@@ -111,7 +111,7 @@ export class SafeEditorComponent implements OnInit {
    */
   private async getData() {
     const metaRes = await firstValueFrom(
-      this.apollo.query<GetResourceMetadataQueryResponse>({
+      this.apollo.query<ResourceQueryResponse>({
         query: GET_RESOURCE_METADATA,
         variables: {
           id: this.settings.resource,
