+++ libs/safe/src/lib/components/widgets/summary-card-settings/graphql/queries.ts
@@ -1,7 +1,6 @@
 import { gql } from 'apollo-angular';
 import { Record } from '../../../../models/record.model';
 import { Resource } from '../../../../models/resource.model';
-import { Layout } from '../../../../models/layout.model';
 
 // === GET RECORD BY ID ===
 /** Graphql request for getting a record by its id */
@@ -31,28 +30,6 @@ export const GET_RECORD_BY_ID = gql`
 /** Model for GetRecordByIdQueryResponse object */
 export interface GetRecordByIdQueryResponse {
   record: Record;
-}
-
-/** Graphql request for getting resource meta date for a grid */
-export const GET_RESOURCE_LAYOUTS = gql`
-  query GetResource($id: ID!) {
-    resource(id: $id) {
-      layouts {
-        id
-        query
-      }
-    }
-  }
-`;
-
-/** Model for GetResourceByIdQueryResponse object */
-export interface GetResourceLayoutsByIdQueryResponse {
-  resource: {
-    layouts: {
-      id: Layout['id'];
-      query: Layout['query'];
-    }[];
-  };
 }
 
 /** Graphql query for getting multiple resources with a cursor */
