+++ libs/safe/src/lib/components/widgets/summary-card/graphql/queries.ts
@@ -1,6 +1,5 @@
 import { gql } from 'apollo-angular';
 import { Record } from '../../../../models/record.model';
-import { Layout } from '../../../../models/layout.model';
 import { Resource } from '../../../../models/resource.model';
 
 // === GET RECORD BY ID ===
@@ -33,28 +32,6 @@ export interface GetRecordByIdQueryResponse {
   record: Record;
 }
 
-/** Graphql request for getting resource meta date for a grid */
-export const GET_RESOURCE_LAYOUTS = gql`
-  query GetResource($id: ID!) {
-    resource(id: $id) {
-      layouts {
-        id
-        query
-      }
-    }
-  }
-`;
-
-/** Model for GetResourceByIdQueryResponse object */
-export interface GetResourceLayoutsByIdQueryResponse {
-  resource: {
-    layouts: {
-      id: Layout['id'];
-      query: Layout['query'];
-    }[];
-  };
-}
-
 /** Graphql request for getting resource metadata */
 export const GET_RESOURCE_METADATA = gql`
   query GetResourceMeta($id: ID!) {
@@ -67,11 +44,6 @@ export const GET_RESOURCE_METADATA = gql`
     }
   }
 `;
-
-/** Model for GetResourceByIdQueryResponse object */
-export interface GetResourceMetadataQueryResponse {
-  resource: Resource;
-}
 
 /** Graphql request for getting resource layout */
 export const GET_LAYOUT = gql`
