+++ libs/safe/src/lib/components/workflow-stepper/components/add-step/add-step.component.spec.ts
@@ -1,31 +1,12 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import {
-  TranslateModule,
-  TranslateService,
-  TranslateFakeLoader,
-  TranslateLoader,
-} from '@ngx-translate/core';
-
 import { SafeAddStepComponent } from './add-step.component';
+import { setupSpecConfig } from '../../../ui/aggregation-builder/pipeline/shared-config/spec/config';
 
 describe('SafeAddStepComponent', () => {
   let component: SafeAddStepComponent;
   let fixture: ComponentFixture<SafeAddStepComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      providers: [TranslateService],
-      declarations: [SafeAddStepComponent],
-      imports: [
-        TranslateModule.forRoot({
-          loader: {
-            provide: TranslateLoader,
-            useClass: TranslateFakeLoader,
-          },
-        }),
-      ],
-    }).compileComponents();
-  });
+  beforeEach(async () => setupSpecConfig(SafeAddStepComponent));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(SafeAddStepComponent);
