+++ libs/safe/src/lib/graphql/queries/public-api.ts
@@ -1,0 +1,1 @@
+export * from './resource-queries';
