+++ libs/safe/src/lib/graphql/queries/resource-queries.ts
@@ -1,8 +1,19 @@
 import { gql } from 'apollo-angular';
-import { Resource } from '@oort-front/safe';
+
+/** Graphql request for getting resource meta date for a grid */
+export const GET_RESOURCE_LAYOUTS = gql`
+  query GetResource($id: ID!) {
+    resource(id: $id) {
+      layouts {
+        id
+        query
+      }
+    }
+  }
+`;
 
 /** Graphql query for getting a resource by its id */
-export const GET_RESOURCE_LAYOUTS = gql`
+export const GET_RESOURCE_LAYOUTS_FULL = gql`
   query GetResourceById($id: ID!, $first: Int, $afterCursor: ID) {
     resource(id: $id) {
       id
@@ -26,8 +37,3 @@ export const GET_RESOURCE_LAYOUTS = gql`
     }
   }
 `;
-
-/** Model for GetResourceByIdQueryResponse object */
-export interface GetResourceByIdQueryResponse {
-  resource: Resource;
-}
